tried V3 dual can(not working)

+++ lexxpluss_apps/src/rosserial_tof.hpp
@@ -50,7 +50,7 @@ public:
     }
 private:
     std_msgs::Float64MultiArray msg;
-    float msg_data[2];
+    double msg_data[2];
     ros::Publisher pub{"/sensor_set/downward", &msg};
 };
